backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pypdf2
requests
